+++ packages/utils/src/arr/includes.ts
@@ -1,0 +1,16 @@
+import type { Suggest, UnknownArray } from '@ylfjuk-core/types';
+
+/**
+ * Checks if an array includes a specific element.
+ *
+ * @note A wrapper around {@link Array.prototype.includes} that allows to check against any type
+ *
+ * @since 0.0.8
+ */
+export const arrIncludes = <
+    const Arr extends UnknownArray,
+    const El extends Suggest<Arr[number]>
+>(
+    arr: Arr,
+    el: El
+): boolean => arr.includes(el);