+++ packages/utils/src/arr/index.ts
@@ -1,0 +1,12 @@
+import { arrIncludes } from './includes';
+
+/**
+ * Utility functions for arrays.
+ *
+ * @since 0.0.8
+ */
+export const Arr = {
+    includes: arrIncludes,
+};
+
+export { arrIncludes };