+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @ylfjuk-core/utils
 
+## 0.0.9
+
+### Patch Changes
+
+#### Features 🆕
+
+- Added
+  - jsonParse | A wrapper around `JSON.parse` that fixes the type
+  - jsonSafeParse | A wrapper around `jsonParse` that returns a `Result` value instead of throwing
+  - stringify | A wrapper around `JSON.stringify` that fixes the type
+  - Json | Utility functions for JSON
+
+## 0.0.8
+
+### Patch Changes
+
+#### Features 🆕
+
+- Added
+  - bool | A wrapper around `Boolean` that narrows the type to exclude falsy values
+  - keysOf | A wrapper around `Object.keys` that retains the type of the keys
+  - entriesOf | A wrapper around `Object.entries` that retains the type of the keys and values
+  - arrIncludes | A wrapper around `Array.prototype.includes` that allows to check against any type
+  - strStartsWith | A wrapper around `String.prototype.startsWith` that narrows the type to a template literal type
+  - strEndsWith | A wrapper around `String.prototype.endsWith` that narrows the type to a template literal type
+  - Arr | Utility functions for arrays
+  - Obj | Utility functions for objects
+  - Str | Utility functions for strings
+
+#### Chores 🧹
+
+- Added `"sideEffects": false` to the `package.json`
+
 ## 0.0.7
 
 ### Patch Changes